Fix duplicate

+++ validation/trans-correctness/phase3/output/location-service/src/main/java/com/example/location/repository/LocationRepository.java
@@ -1,8 +1,9 @@
+
 package com.example.location.repository;
 
 import com.example.location.domain.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<Location,Integer > {
+public interface LocationRepository extends JpaRepository<Location, Long> {
     public Boolean existsByPortCode(Integer portCode);
 }
